requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
